+++ pypeit/core/skysub.py
@@ -14,134 +14,14 @@ from pypeit import utils
 from pypeit import debugger
 from pypeit.core import pixels
 from pypeit.core import extract
+from matplotlib import pyplot as plt
 
 from scipy.special import ndtr
 
 
 
-# ToDO Fix masking logic. This code should also take an ivar for consistency with rest of extraction
-def bg_subtraction_slit(slit, slitpix, edge_mask, sciframe, varframe, tilts,
-                        bpm=None, crmask=None, tracemask=None, bsp=0.6, sigrej=3., POS_MASK=True,
-                        PLOT_FIT=False):
-    """
-    Perform sky subtraction on an input slit
-
-    Parameters
-    ----------
-    slit : int
-      Slit number; indexed 1, 2,
-    slitpix : ndarray
-      Specifies pixels in the slits
-    edgemask : ndarray
-      Mask edges of the slit
-    sciframe : ndarray
-      science frame
-    varframe : ndarray
-      Variance array
-    tilts : ndarray
-      Tilts of the wavelengths
-    bpm : ndarray, optional
-      Bad pixel mask
-    crmask : ndarray
-      Cosmic ray mask
-    tracemask : ndarray
-      Object mask
-    bsp : float
-      Break point spacing
-    sigrej : float
-      rejection
-
-    Returns
-    -------
-    bgframe : ndarray
-      Sky background image
-
-    """
-
-    # Init
-    bgframe = np.zeros_like(sciframe)
-    ivar = utils.calc_ivar(varframe)
-    ny = sciframe.shape[0]
-    piximg = tilts * (ny-1)
-
-    #
-    ordpix = slitpix.copy()
-    # Masks
-    if bpm is not None:
-        ordpix *= 1-bpm.astype(np.int)
-    if crmask is not None:
-        ordpix *= 1-crmask.astype(np.int)
-    if tracemask is not None:
-        ordpix *= (1-tracemask.astype(np.int))
-
-    # Sky pixels for fitting
-    fit_sky = (ordpix == slit) & (ivar > 0.) & (~edge_mask)
-    isrt = np.argsort(piximg[fit_sky])
-    wsky = piximg[fit_sky][isrt]
-    sky = sciframe[fit_sky][isrt]
-    sky_ivar = ivar[fit_sky][isrt]
-
-    # All for evaluation
-    all_slit = (slitpix == slit) & (~edge_mask)
-
-    # Restrict fit to positive pixels only and mask out large outliers via a pre-fit to the log
-    if (POS_MASK==True):
-        pos_sky = (sky > 1.0) & (sky_ivar > 0.)
-        if np.sum(pos_sky) > ny:
-            lsky = np.log(sky[pos_sky])
-            lsky_ivar = lsky * 0. + 0.1
-
-            # Init bspline to get the sky breakpoints (kludgy)
-            tmp = pydl.bspline(wsky[pos_sky], nord=4, bkspace=bsp)
-
-            #skybkpt = bspline_bkpts(wsky[pos_sky], nord=4, bkspace=bsp $
-            #, / silent)
-            if False:
-                plt.clf()
-                ax = plt.gca()
-                ax.scatter(wsky[pos_sky], lsky)
-                #ax.scatter(wsky[~full_out], sky[~full_out], color='red')
-                #ax.plot(wsky, yfit, color='green')
-                plt.show()
-                #debugger.set_trace()
-            lskyset, outmask, lsky_fit, red_chi = utils.bspline_profile(
-                wsky[pos_sky], lsky, lsky_ivar, np.ones_like(lsky),
-                fullbkpt = tmp.breakpoints, upper=sigrej, lower=sigrej,
-                kwargs_reject={'groupbadpix':True})
-            res = (sky[pos_sky] - np.exp(lsky_fit)) * np.sqrt(sky_ivar[pos_sky])
-            lmask = (res < 5.0) & (res > -4.0)
-            sky_ivar[pos_sky] = sky_ivar[pos_sky] * lmask
-
-    # Full fit now
-    full_bspline = pydl.bspline(wsky, nord=4, bkspace=bsp)
-    skyset, full_out, yfit, _ = utils.bspline_profile(
-        wsky, sky, sky_ivar, np.ones_like(sky),
-        fullbkpt=full_bspline.breakpoints,
-        upper=sigrej, lower=sigrej, kwargs_reject={'groupbadpix':True, 'maxrej': 10})
-    # Evaluate and save
-    bgframe[all_slit] = skyset.value(piximg[all_slit])[0] #, skyset)
-
-    # Debugging/checking
-    if PLOT_FIT:
-        goodbk = skyset.mask
-        yfit_bkpt = np.interp(skyset.breakpoints[goodbk], wsky,yfit)
-        plt.clf()
-        ax = plt.gca()
-        was_fit = (sky_ivar > 0.0)
-        was_fit_and_masked = (was_fit == True) & (full_out == False)
-        ax.plot(wsky[was_fit], sky[was_fit], color='k', marker='o', markersize=0.4, mfc='k', fillstyle='full', linestyle='None')
-        ax.plot(wsky[was_fit_and_masked], sky[was_fit_and_masked], color='red', marker='+', markersize=1.5, mfc='red', fillstyle='full', linestyle='None')
-        ax.plot(wsky, yfit, color='cornflowerblue')
-        ax.plot(skyset.breakpoints[goodbk], yfit_bkpt, color='lawngreen', marker='o', markersize=2.0, mfc='lawngreen', fillstyle='full', linestyle='None')
-        ax.set_ylim((0.99*yfit.min(),1.01*yfit.max()))
-        plt.show()
-
-    # Return
-    return bgframe
-
-
 def global_skysub(image, ivar, tilts, thismask, slit_left, slit_righ, inmask = None, bsp=0.6, sigrej=3., trim_edg = (3,3),
-                  pos_mask=True, show_fit=False):
+                  pos_mask=True, show_fit=False, no_poly=False, npoly = None):
     """
     Perform global sky subtraction on an input slit
 
@@ -208,54 +88,98 @@ def global_skysub(image, ivar, tilts, thismask, slit_left, slit_righ, inmask = N
 
     ximg, edgmask = pixels.ximg_and_edgemask(slit_left, slit_righ, thismask, trim_edg=trim_edg)
 
+
     # Init
-    nspec = image.shape[0]
+    (nspec, nspat) = image.shape
     piximg = tilts * (nspec-1)
     if inmask is None:
         inmask = np.copy(thismask)
 
-    # Sky pixels for fitting
-    fit_sky = (thismask == True) & (ivar > 0.0) & (inmask == True) & (edgmask == False)
-    isrt = np.argsort(piximg[fit_sky])
-    wsky = piximg[fit_sky][isrt]
-    sky = image[fit_sky][isrt]
-    sky_ivar = ivar[fit_sky][isrt]
 
-    # ToDo Add in a polynomial profile basis here to improve the global sky subtraction?
-    # Restrict fit to positive pixels only and mask out large outliers via a pre-fit to the log
+    # Sky pixels for fitting
+    inmask_in = (thismask == True) & (ivar > 0.0) & (inmask == True) & (edgmask == False)
+    isrt = np.argsort(piximg[thismask])
+    pix = piximg[thismask][isrt]
+    sky = image[thismask][isrt]
+    sky_ivar = ivar[thismask][isrt]
+    ximg_fit = ximg[thismask][isrt]
+    inmask_fit = inmask_in[thismask][isrt]
+    #spatial = spatial_img[fit_sky][isrt]
+
+    # Restrict fit to positive pixels only and mask out large outliers via a pre-fit to the log.
     if (pos_mask is True):
         pos_sky = (sky > 1.0) & (sky_ivar > 0.0)
         if np.sum(pos_sky) > nspec:
             lsky = np.log(sky[pos_sky])
-            lsky_ivar = np.full(lsky.shape, 0.1)
+            lsky_ivar = inmask_fit[pos_sky].astype(float)/3.0** 2  # set errors to just be 3.0 in the log
+            #lsky_ivar = np.full(lsky.shape, 0.1)
             # Init bspline to get the sky breakpoints (kludgy)
-            tmp = pydl.bspline(wsky[pos_sky], nord=4, bkspace=bsp)
-            lskyset, outmask, lsky_fit, red_chi = utils.bspline_profile(wsky[pos_sky], lsky, lsky_ivar, np.ones_like(lsky),
-                                                                        fullbkpt = tmp.breakpoints, upper=sigrej, lower=sigrej,
+            #tmp = pydl.bspline(wsky[pos_sky], nord=4, bkspace=bsp)
+            lskyset, outmask, lsky_fit, red_chi = utils.bspline_profile(pix[pos_sky], lsky, lsky_ivar, np.ones_like(lsky),
+                                                                        inmask = inmask_fit[pos_sky],
+                                                                        upper=sigrej, lower=sigrej,
+                                                                        kwargs_bspline={'bkspace':bsp},
                                                                         kwargs_reject={'groupbadpix': True, 'maxrej': 10})
             res = (sky[pos_sky] - np.exp(lsky_fit)) * np.sqrt(sky_ivar[pos_sky])
             lmask = (res < 5.0) & (res > -4.0)
             sky_ivar[pos_sky] = sky_ivar[pos_sky] * lmask
+            inmask_fit[pos_sky]=(sky_ivar[pos_sky] > 0.0) & lmask
 
+    if no_poly:
+        poly_basis= np.ones_like(sky)
+    else:
+        npercol = np.fmax(np.floor(np.sum(thismask) / nspec), 1.0)
+        # Demand at least 10 pixels per row (on average) per degree of the polynomial
+        if npoly is None:
+            #npoly_in = 7
+            #npoly = np.fmax(np.fmin(npoly_in, (np.ceil(npercol / 10.)).astype(int)), 1)
+            if npercol > 100:
+                npoly = 3
+            elif npercol > 40:
+                npoly = 2
+            else:
+                npoly = 1
+        poly_basis = pydl.flegendre(2.0*ximg_fit - 1.0, npoly).T
 
     # Full fit now
-    full_bspline = pydl.bspline(wsky, nord=4, bkspace=bsp)
-    skyset, outmask, yfit, _ = utils.bspline_profile(wsky, sky, sky_ivar, np.ones_like(sky),
-                                                       fullbkpt=full_bspline.breakpoints,upper=sigrej, lower=sigrej,
-                                                       kwargs_reject={'groupbadpix':True, 'maxrej': 10})
+    #full_bspline = pydl.bspline(wsky, nord=4, bkspace=bsp, npoly = npoly)
+    #skyset, outmask, yfit, _ = utils.bspline_profile(wsky, sky, sky_ivar, poly_basis,
+    #                                                   fullbkpt=full_bspline.breakpoints,upper=sigrej, lower=sigrej,
+    #                                                   kwargs_reject={'groupbadpix':True, 'maxrej': 10})
+
+
+    # Perform the full fit now
+    skyset, outmask, yfit, _ = utils.bspline_profile(pix, sky, sky_ivar,poly_basis,inmask = inmask_fit, nord = 4,
+                                                               upper=sigrej, lower=sigrej,
+                                                               kwargs_bspline = {'bkspace':bsp},
+                                                               kwargs_reject={'groupbadpix':True, 'maxrej': 10})
+
+    sky_frame = np.zeros_like(image)
+    ythis = np.zeros_like(yfit)
+    ythis[isrt] = yfit
+    sky_frame[thismask] = ythis
+
+    #skyset.funcname ='legendre'
+    #skyset.xmin = spat_min
+    #skyset.xmax = spat_max
+
     # Evaluate and save
-    bgframe, _ = skyset.value(piximg[thismask])
+    #bgframe, _ = skyset.value(piximg[thismask],x2=spatial_img[thismask])
 
     # Debugging/checking
+
+    # ToDo This QA ceases to make sense I think for 2-d fits. I need to think about what the best QA would be here, but I think
+    # probably looking at residuals as a function of spectral and spatial position like in the flat fielding code.
     if show_fit:
         goodbk = skyset.mask
-        yfit_bkpt = np.interp(skyset.breakpoints[goodbk], wsky,yfit)
+        # This is approximate
+        yfit_bkpt = np.interp(skyset.breakpoints[goodbk], pix,yfit)
         plt.clf()
         ax = plt.gca()
-        was_fit_and_masked = (outmask == False)
-        ax.plot(wsky, sky, color='k', marker='o', markersize=0.4, mfc='k', fillstyle='full', linestyle='None')
-        ax.plot(wsky[was_fit_and_masked], sky[was_fit_and_masked], color='red', marker='+', markersize=1.5, mfc='red', fillstyle='full', linestyle='None')
-        ax.plot(wsky, yfit, color='cornflowerblue')
+        was_fit_and_masked = inmask_fit & ~outmask
+        ax.plot(pix, sky, color='k', marker='o', markersize=0.4, mfc='k', fillstyle='full', linestyle='None')
+        ax.plot(pix[was_fit_and_masked], sky[was_fit_and_masked], color='red', marker='+', markersize=1.5, mfc='red', fillstyle='full', linestyle='None')
+        ax.plot(pix, yfit, color='cornflowerblue')
         ax.plot(skyset.breakpoints[goodbk], yfit_bkpt, color='lawngreen', marker='o', markersize=4.0, mfc='lawngreen', fillstyle='full', linestyle='None')
         ax.set_ylim((0.99*yfit.min(),1.01*yfit.max()))
         plt.show()
@@ -264,7 +188,7 @@ def global_skysub(image, ivar, tilts, thismask, slit_left, slit_righ, inmask = N
     # ToDO worth thinking about whether we want to return a mask here. It makese no sense to return outmask
     # in its present form though since that does not refer to the whole image.
     # return bgframe, outmask
-    return bgframe
+    return ythis
 
 
 
@@ -299,7 +223,9 @@ def skyoptimal(wave,data,ivar, oprof, sortpix, sigrej = 3.0, npoly = 1, spatial 
     good = good[wave[good].argsort()]
     relative, = np.where(relative_mask[good])
 
-    (sset1, outmask_good1, yfit1, red_chi1) = utils.bspline_profile(wave[good], data[good], ivar[good], profile_basis[good, :],
+
+
+    sset1, outmask_good1, yfit1, red_chi1 = utils.bspline_profile(wave[good], data[good], ivar[good], profile_basis[good, :],
                                                               fullbkpt=fullbkpt, upper=sigrej, lower=sigrej,
                                                               relative=relative,
                                                               kwargs_reject={'groupbadpix': True, 'maxrej': 5})
@@ -313,7 +239,7 @@ def skyoptimal(wave,data,ivar, oprof, sortpix, sigrej = 3.0, npoly = 1, spatial 
     msgs.info('2nd round....')
     msgs.info('Iter     Chi^2     Rejected Pts')
 
-    (sset, outmask_good, yfit, red_chi) = utils.bspline_profile(wave[good], data[good], ivar[good] * mask1,
+    sset, outmask_good, yfit, red_chi = utils.bspline_profile(wave[good], data[good], ivar[good] * mask1,
                                                           profile_basis[good, :],
                                                           fullbkpt=fullbkpt, upper=sigrej, lower=sigrej,
                                                           relative=relative,
@@ -450,19 +376,20 @@ def local_skysub_extract(sciimg, sciivar, tilts, waveimg, global_sky, rn2_img, t
             npoly = 1
         obj_profiles = np.zeros((nspec, nspat, objwork), dtype=float)
         sigrej_eff = sigrej
-        for iiter in range(1, niter):
-            msgs.info("Iteration # " + "{:2d}".format(iiter) + " of " + "{:2d}".format(niter))
+        proc_list = [] # List of processes for interactive plotting, these are terminated at the end of each iteration sequence
+        for iiter in range(1, niter + 1):
+            msgs.info('--------------------------REDUCING: Iteration # ' + '{:2d}'.format(iiter) + ' of ' +
+                      '{:2d}'.format(niter) + '---------------------------------------------------')
             img_minsky = sciimg - skyimage
             for ii in range(objwork):
                 iobj = group[ii]
                 if iiter == 1:
                     # If this is the first iteration, print status message. Initiate profile fitting with a simple
                     # boxcar extraction.
-                    msgs.info("-------------------REDUCING-------------------")
-                    msgs.info("Fitting profile for obj #: " + "{:}".format(sobjs[iobj].objid) + " of {:}".format(nobj))
-                    msgs.info(
-                        "At x = {:5.2f}".format(sobjs[iobj].spat_pixpos) + " on slit # {:}".format(sobjs[iobj].slitid))
-                    msgs.info("----------------------------------------------")
+                    msgs.info("----------------------------------- PROFILE FITTING --------------------------------------------------------")
+                    msgs.info("Fitting profile for obj # " + "{:}".format(sobjs[iobj].objid) + " of {:}".format(nobj))
+                    msgs.info("At x = {:5.2f}".format(sobjs[iobj].spat_pixpos) + " on slit # {:}".format(sobjs[iobj].slitid))
+                    msgs.info("------------------------------------------------------------------------------------------------------------")
                     flux = extract.extract_boxcar(img_minsky * outmask, sobjs[iobj].trace_spat, box_rad,
                                           ycen=sobjs[iobj].trace_spec)
                     mvarimg = 1.0 / (modelivar + (modelivar == 0))
@@ -490,15 +417,19 @@ def local_skysub_extract(sciimg, sciivar, tilts, waveimg, global_sky, rn2_img, t
                         fluxivar = sobjs[iobj].optimal['COUNTS_IVAR']
                         wave = sobjs[iobj].optimal['WAVE']
 
+                obj_string = 'obj # {:}'.format(sobjs[iobj].objid) + ' on slit # {:}'.format(sobjs[iobj].slitid) + ', iter # {:}'.format(iiter) + ':'
                 if wave.any():
-                    (profile_model, xnew, fwhmfit, med_sn2) = extract.fit_profile(img_minsky[ipix], (modelivar * outmask)[ipix],
+                    (profile_model, xnew, fwhmfit, med_sn2, show_proc) = extract.fit_profile(img_minsky[ipix], (modelivar * outmask)[ipix],
                                                                           waveimg[ipix],
                                                                           sobjs[iobj].trace_spat - mincol,
                                                                           wave, flux, fluxivar,
                                                                           thisfwhm=sobjs[iobj].fwhm,
                                                                           hwidth=sobjs[iobj].maskwidth,
                                                                           prof_nsigma=sobjs[iobj].prof_nsigma,
-                                                                          sn_gauss=sn_gauss, show_profile=show_profile)
+                                                                          sn_gauss=sn_gauss, obj_string = obj_string,
+                                                                          show_profile=show_profile)
+                    proc_list.append(show_proc)
+
                     # Update the object profile and the fwhm and mask parameters
                     obj_profiles[ipix[0], ipix[1], ii] = profile_model
                     sobjs[iobj].trace_spat = xnew + mincol
@@ -599,6 +530,14 @@ def local_skysub_extract(sciimg, sciivar, tilts, waveimg, global_sky, rn2_img, t
                 msgs.warn('ERROR: Bspline sky subtraction failed after 4 iterations of bkpt spacing')
                 msgs.warn('       Moving on......')
                 obj_profiles = np.zeros_like(obj_profiles)
+                # Just replace with the global sky
+                skyimage.flat[isub] = global_sky.flat[isub]
+
+        # Now that iterations are complete, clear the windows if show_profile was set
+        if show_profile:
+            for proc in proc_list:
+                proc.terminate()
+                proc.join()
 
         # Now that the iterations of profile fitting and sky subtraction are completed,
         # loop over the objwork objects in this grouping and perform the final extractions.
@@ -667,189 +606,14 @@ def local_skysub_extract(sciimg, sciivar, tilts, waveimg, global_sky, rn2_img, t
         canvas_list = points_mask + points_omask + text_mask + text_omask
         canvas.add('constructedcanvas', canvas_list)
 
+    # Clean up any profile plots
+#    if show_profile:
+#        try:
+#            show_proc.terminate()
+#            show_proc.join()
+#        except:
+#            pass
 
     return (skyimage[thismask], objimage[thismask], modelivar[thismask], outmask[thismask])
-
-def order_pixels(pixlocn, lord, rord):
-    """
-    Based on physical pixel locations, determine which pixels are within the orders
-    """
-
-    sz_x, sz_y, _ = pixlocn.shape
-    sz_o = lord.shape[1]
-
-    outfr = np.zeros((sz_x, sz_y), dtype=int)
-
-    for y in range(sz_y):
-        for o in range(sz_o):
-            indx = (lord[:,o] < rord[:,o]) & (pixlocn[:,y,1] > lord[:,o]) \
-                   & (pixlocn[:,y,1] < rord[:,o])
-            indx |= ( (lord[:,o] > rord[:,o]) & (pixlocn[:,y,1] < lord[:,o])
-                      & (pixlocn[:,y,1] > rord[:,o]) )
-            if np.any(indx):
-                # Only assign a single order to a given pixel
-                outfr[indx,y] = o+1
-                break
-
-    return outfr
-
-
-# This code is deprecated and replaced by bg_subtraction_slit
-def orig_bg_subtraction_slit(tslits_dict, pixlocn,
-                        slit, tilts, sciframe, varframe, bpix, crpix,
-                        settings,
-                        tracemask=None,
-                        rejsigma=3.0, maskval=-999999.9,
-                        method='bspline', debug=False):
-    """ Extract a science target and background flux
-    :param slf:
-    :param sciframe:
-    :param varframe:
-    :return:
-    """
-    # Unpack tslits
-    lordloc = tslits_dict['lcen']
-    rordloc = tslits_dict['rcen']
-    slitpix = tslits_dict['slitpix']
-    # Init
-    bgframe = np.zeros_like(sciframe)
-
-    # Begin the algorithm
-    # Find which pixels are within the order edges
-    msgs.info("Identifying pixels within each order")
-    ordpix = order_pixels(pixlocn, lordloc*0.95+rordloc*0.05, lordloc*0.05+rordloc*0.95)
-
-    msgs.info("Applying bad pixel mask")
-    ordpix *= (1-bpix.astype(np.int)) * (1-crpix.astype(np.int))
-    if tracemask is not None: ordpix *= (1-tracemask.astype(np.int))
-
-    # Construct an array of pixels to be fit with a spline
-    whord = np.where(ordpix == slit+1)
-    xvpix  = tilts[whord]
-    scipix = sciframe[whord]
-    xargsrt = np.argsort(xvpix, kind='mergesort')
-    sxvpix  = xvpix[xargsrt]
-    sscipix = scipix[xargsrt]
-
-    # Reject deviant pixels -- step through every 1.0/sciframe.shape[0] in sxvpix and reject significantly deviant pixels
-    edges = np.linspace(min(0.0,np.min(sxvpix)),max(1.0,np.max(sxvpix)),sciframe.shape[0])
-    fitcls = np.zeros(sciframe.shape[0])
-
-    # Identify science target
-    maskpix = np.zeros(sxvpix.size)
-    msgs.info("Identifying pixels containing the science target")
-    msgs.work("Speed up this step with multi-processing")
-    for i in range(sciframe.shape[0]-1):
-        wpix = np.where((sxvpix>=edges[i]) & (sxvpix<=edges[i+1]))
-        if (wpix[0].size>5):
-            txpix = sxvpix[wpix]
-            typix = sscipix[wpix]
-            msk, cf = utils.robust_polyfit(txpix, typix, 0, sigma=rejsigma)
-            maskpix[wpix] = msk
-            #fitcls[i] = cf[0]
-            wgd=np.where(msk == 0)
-            szt = np.size(wgd[0])
-            if szt > 8:
-                fitcls[i] = np.mean(typix[wgd][szt//2-3:szt//2+4]) # Average the 7 middle pixels
-                #fitcls[i] = np.mean(np.random.shuffle(typix[wgd])[:5]) # Average the 5 random pixels
-            else:
-                fitcls[i] = cf[0]
-    '''
-    else:
-        debugger.set_trace()
-        msgs.work("Speed up this step in cython")
-        for i in range(sciframe.shape[0]-1):
-            wpix = np.where((sxvpix >= edges[i]) & (sxvpix <= edges[i+1]))
-            typix = sscipix[wpix]
-            szt = typix.size
-            if szt > 8:
-                fitcls[i] = np.mean(typix[szt//2-3:szt//2+4])  # Average the 7 middle pixels
-            elif szt != 0:
-                fitcls[i] = np.mean(typix)
-            else:
-                fitcls[i] = 0.0
-        # Trace the sky lines to get a better estimate of the tilts
-        scicopy = sciframe.copy()
-        scicopy[np.where(ordpix == slit)] = maskval
-        scitilts, _ = artrace.model_tilt(det, scicopy, guesstilts=tilts.copy(),
-                                         censpec=fitcls, maskval=maskval, plotQA=True)
-        xvpix  = scitilts[whord]
-        scipix = sciframe[whord]
-        varpix = varframe[whord]
-        mskpix = tracemask[whord]
-        xargsrt = np.argsort(xvpix, kind='mergesort')
-        sxvpix  = xvpix[xargsrt]
-        sscipix = scipix[xargsrt]
-        svarpix = varpix[xargsrt]
-        maskpix = mskpix[xargsrt]
-    '''
-
-    # Check the mask is reasonable
-    scimask = sciframe.copy()
-    rxargsrt = np.argsort(xargsrt, kind='mergesort')
-    scimask[whord] *= (1.0-maskpix)[rxargsrt]
-
-    # Now trace the sky lines to get a better estimate of the spectral tilt during the observations
-    scifrcp = scimask.copy()
-    scifrcp[whord] += (maskval*maskpix)[rxargsrt]
-    scifrcp[np.where(ordpix != slit+1)] = maskval
-
-    # Check tilts? -- Can also be error in flat fielding or slit illumination
-    if False:
-        idx = 1893
-        plt.clf()
-        ax = plt.gca()
-        ax.scatter(tilts[idx-2,:], scifrcp[idx-2,:], color='green')
-        ax.scatter(tilts[idx-1,:], scifrcp[idx-1,:], color='blue')
-        ax.scatter(tilts[idx,:], scifrcp[idx,:], color='red')
-        ax.scatter(tilts[idx+1,:], scifrcp[idx+1,:], color='orange')
-        ax.set_ylim(0., 3000)
-        plt.show()
-        debugger.set_trace()
-    #
-    msgs.info("Fitting sky background spectrum")
-    if method == 'bspline':
-        msgs.info("Using bspline sky subtraction")
-        gdp = (scifrcp != maskval) & (ordpix == slit+1) & (varframe > 0.)
-        srt = np.argsort(tilts[gdp])
-        #bspl = utils.func_fit(tilts[gdp][srt], scifrcp[gdp][srt], 'bspline', 3,
-        #                        **settings.argflag['reduce']['skysub']['bspline'])
-        ivar = utils.calc_ivar(varframe)
-        mask, bspl = utils.robust_polyfit(tilts[gdp][srt], scifrcp[gdp][srt], 3,
-                                            function='bspline',
-                                            weights=np.sqrt(ivar)[gdp][srt],
-                                            sigma=5., maxone=False,
-                                            bspline_par=settings['skysub']['bspline'])
-        # Just those in the slit
-        in_slit = np.where(slitpix == slit+1)
-        bgf_flat = utils.func_val(bspl, tilts[in_slit].flatten(), 'bspline')
-        #bgframe = bgf_flat.reshape(tilts.shape)
-        bgframe[in_slit] = bgf_flat
-        if debug:
-            plt_bspline_sky(tilts, scifrcp, bgf_flat, in_slit, gdp)
-            debugger.set_trace()
-    else:
-        msgs.error('Not ready for this method for skysub {:s}'.format(method))
-
-    if np.sum(np.isnan(bgframe)) > 0:
-        msgs.warn("NAN in bgframe.  Replacing with 0")
-        bad = np.isnan(bgframe)
-        bgframe[bad] = 0.
-
-    # Plot to make sure that the result is good
-    return bgframe
-
-
-
-def plt_bspline_sky(tilts, scifrcp, bgf_flat, inslit, gdp):
-    # Setup
-    srt = np.argsort(tilts[inslit].flatten())
-    # Plot
-    plt.close()
-    plt.clf()
-    ax = plt.gca()
-    ax.scatter(tilts[gdp]*tilts.shape[0], scifrcp[gdp], marker='o')
-    ax.plot(tilts[inslit].flatten()[srt]*tilts.shape[0], bgf_flat[srt], 'r-')
-    plt.show()